+++ lib/libesp32/berry/berry_port/be_byteslib.py
@@ -1098,6 +1098,171 @@ def m_addfloat(vm):
     be_api.be_raise(vm, "type_error", "operands must be int or float")
     return be_api.be_returnnilvalue(vm)
 
+# static int m_getbits(bvm *vm)
+# {
+#     int argc = be_top(vm);
+#     buf_impl attr = bytes_check_data(vm, 0);
+#     check_ptr(vm, &attr);
+#     if (argc >= 3 && be_isint(vm, 2) && be_isint(vm, 3)) {
+#         int32_t offset_bits = be_toint(vm, 2);
+#         int32_t len_bits = be_toint(vm, 3);
+#         if (len_bits < -32 || len_bits > 32) {
+#             be_raise(vm, "value_error", "length in bits must be between 0 and 32");
+#         }
+#         if (len_bits == 0) { be_return_nil(vm); }
+#         bbool big_endian = bfalse;
+#         if (len_bits < 0) { big_endian = btrue; len_bits = -len_bits; }
+#         uint32_t ret = 0;
+#         int32_t offset_bytes = offset_bits >> 3;
+#         offset_bits = offset_bits & 7;
+#         int bit_shift = 0;
+#         while (len_bits > 0) {
+#             int block_bits = 8 - offset_bits;
+#             if (block_bits > len_bits) { block_bits = len_bits; }
+#             uint8_t byte_val = buf_get1(&attr, offset_bytes);
+#             uint32_t mask_val = (1u << block_bits) - 1;
+#             if (big_endian) {
+#                 bit_shift = 8 - offset_bits - block_bits;
+#                 ret = (ret << block_bits) | ((byte_val >> bit_shift) & mask_val);
+#             } else {
+#                 ret |= (uint32_t)(((byte_val >> offset_bits) & mask_val) << bit_shift);
+#                 bit_shift += block_bits;
+#             }
+#             len_bits -= block_bits;
+#             offset_bits = 0;
+#             offset_bytes += 1;
+#         }
+#         be_pop(vm, argc - 1);
+#         be_pushint(vm, (bint) ret);
+#         be_return(vm);
+#     }
+#     be_raise(vm, "type_error", "operands must be int");
+#     be_return_nil(vm);
+# }
+def m_getbits(vm):
+    be_api = _lazy_be_api()
+    argc = be_api.be_top(vm)
+    attr = bytes_check_data(vm, 0)
+    check_ptr(vm, attr)
+    if argc >= 3 and be_api.be_isint(vm, 2) and be_api.be_isint(vm, 3):
+        offset_bits = be_api.be_toint(vm, 2)
+        len_bits = be_api.be_toint(vm, 3)
+        if len_bits < -32 or len_bits > 32:
+            be_api.be_raise(vm, "value_error", "length in bits must be between 0 and 32")
+        if len_bits == 0:
+            return be_api.be_returnnilvalue(vm)
+        big_endian = len_bits < 0
+        if big_endian:
+            len_bits = -len_bits
+        ret = 0
+        offset_bytes = offset_bits >> 3
+        offset_bits = offset_bits & 7
+        bit_shift = 0                       # bit number to write to
+        while len_bits > 0:
+            block_bits = 8 - offset_bits    # how many bits to read in the current block (block = byte)
+            if block_bits > len_bits:
+                block_bits = len_bits
+            byte_val = buf_get1(attr, offset_bytes)
+            mask_val = (1 << block_bits) - 1
+            if big_endian:
+                bit_shift = 8 - offset_bits - block_bits    # non-zero only on the last partial byte
+                ret = ((ret << block_bits) | ((byte_val >> bit_shift) & mask_val)) & 0xFFFFFFFF  # shift ret left to make space
+            else:
+                ret |= ((byte_val >> offset_bits) & mask_val) << bit_shift  # append to the left of ret
+                bit_shift += block_bits
+            # move the input window
+            len_bits -= block_bits
+            offset_bits = 0                 # start at full next byte
+            offset_bytes += 1
+        be_api.be_pop(vm, argc - 1)
+        be_api.be_pushint(vm, ret & 0xFFFFFFFF)
+        return be_api.be_returnvalue(vm)
+    be_api.be_raise(vm, "type_error", "operands must be int")
+    return be_api.be_returnnilvalue(vm)
+
+# static int m_setbits(bvm *vm)
+# {
+#     int argc = be_top(vm);
+#     buf_impl attr = bytes_check_data(vm, 0);
+#     check_ptr_modifiable(vm, &attr);
+#     if (argc >= 4 && be_isint(vm, 2) && be_isint(vm, 3) && (be_isint(vm, 4) || be_isbool(vm, 4))) {
+#         int32_t offset_bits = be_toint(vm, 2);
+#         int32_t len_bits = be_toint(vm, 3);
+#         uint32_t val = be_isint(vm, 4) ? (uint32_t)be_toint(vm, 4) : (be_tobool(vm, 4) ? 1u : 0u);
+#         if (len_bits < -32 || len_bits > 32) {
+#             be_raise(vm, "value_error", "length in bits must be between -32 and 32");
+#         }
+#         bbool big_endian = bfalse;
+#         if (len_bits < 0) { big_endian = btrue; len_bits = -len_bits; }
+#         int32_t offset_bytes = offset_bits >> 3;
+#         offset_bits = offset_bits & 7;
+#         while (len_bits > 0) {
+#             int block_bits = 8 - offset_bits;
+#             if (block_bits > len_bits) { block_bits = len_bits; }
+#             uint32_t mask_val = (1u << block_bits) - 1;
+#             len_bits -= block_bits;
+#             uint32_t extracted;
+#             if (big_endian) {
+#                 extracted = (val >> len_bits) & mask_val;
+#                 offset_bits = 8 - offset_bits - block_bits;
+#             } else {
+#                 extracted = val & mask_val;
+#                 val >>= block_bits;
+#             }
+#             uint8_t cur = buf_get1(&attr, offset_bytes);
+#             buf_set1(&attr, offset_bytes,
+#                      (uint8_t)((cur & ((mask_val << offset_bits) ^ 0xFF)) | (extracted << offset_bits)));
+#             offset_bits = 0;
+#             offset_bytes += 1;
+#         }
+#         be_pop(vm, argc - 1);
+#         be_return(vm);
+#     }
+#     be_raise(vm, "type_error", "operands must be int");
+#     be_return_nil(vm);
+# }
+def m_setbits(vm):
+    be_api = _lazy_be_api()
+    argc = be_api.be_top(vm)
+    attr = bytes_check_data(vm, 0)
+    check_ptr_modifiable(vm, attr)
+    if (argc >= 4 and be_api.be_isint(vm, 2) and be_api.be_isint(vm, 3)
+            and (be_api.be_isint(vm, 4) or be_api.be_isbool(vm, 4))):
+        offset_bits = be_api.be_toint(vm, 2)
+        len_bits = be_api.be_toint(vm, 3)
+        if be_api.be_isint(vm, 4):
+            val = be_api.be_toint(vm, 4) & 0xFFFFFFFF
+        else:
+            val = 1 if be_api.be_tobool(vm, 4) else 0
+        if len_bits < -32 or len_bits > 32:
+            be_api.be_raise(vm, "value_error", "length in bits must be between -32 and 32")
+        big_endian = len_bits < 0
+        if big_endian:
+            len_bits = -len_bits
+        offset_bytes = offset_bits >> 3
+        offset_bits = offset_bits & 7
+        while len_bits > 0:
+            block_bits = 8 - offset_bits    # how many bits to write in the current block (block = byte)
+            if block_bits > len_bits:
+                block_bits = len_bits
+            mask_val = (1 << block_bits) - 1
+            len_bits -= block_bits
+            if big_endian:
+                extracted = (val >> len_bits) & mask_val
+                offset_bits = 8 - offset_bits - block_bits  # non-zero only on the last partial byte
+            else:
+                extracted = val & mask_val
+                val >>= block_bits
+            cur = buf_get1(attr, offset_bytes)
+            buf_set1(attr, offset_bytes,
+                     (cur & (((mask_val << offset_bits) ^ 0xFF) & 0xFF)) | (extracted << offset_bits))
+            offset_bits = 0                 # start at full next byte
+            offset_bytes += 1
+        be_api.be_pop(vm, argc - 1)         # leave self on top of stack
+        return be_api.be_returnvalue(vm)    # return self
+    be_api.be_raise(vm, "type_error", "operands must be int")
+    return be_api.be_returnnilvalue(vm)
+
 # static int m_setbytes(bvm *vm) { ... }
 def m_setbytes(vm):
     be_api = _lazy_be_api()
@@ -1639,5 +1804,7 @@ def be_load_byteslib(vm):
         ("..", m_connect),
         ("==", m_equal),
         ("!=", m_nequal),
+        ("getbits", m_getbits),
+        ("setbits", m_setbits),
     ]
     be_api.be_regclass(vm, "bytes", members)