+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 
 ## [15.5.0.1]
 ### Added
+- Berry `bytes` methods `setbits`/`getbits` transposed to native and support for big endian
 
 ### Breaking Changed